+++ src/core/Vault.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
+
 import {Ownable2Step} from "@openzeppelin/contracts/access/Ownable2Step.sol";
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {ReentrancyGuard} from "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
@@ -23,26 +24,20 @@ interface IVault {
 
 contract Vault is ERC4626, ERC20Votes, Ownable2Step, ReentrancyGuard {
     using SafeERC20 for IERC20;
-    uint constant NUMBER_OF_DEAD_SHARES = 1000;
-    address constant DEAD_ADDRESS =
-        address(0x000000000000000000000000000000000000dEaD);
-    uint public interestRatePerSecond = InterestRate.INTEREST_RATE_15;
-    uint public lentAmountStored;
-    uint public lastUpdated;
+
+    uint256 constant NUMBER_OF_DEAD_SHARES = 1000;
+    address constant DEAD_ADDRESS = address(0x000000000000000000000000000000000000dEaD);
+    uint256 public interestRatePerSecond = InterestRate.INTEREST_RATE_15;
+    uint256 public lentAmountStored;
+    uint256 public lastUpdated;
     address public config;
     mapping(address => bool) public isWhitelisted;
     address public governor;
 
-    constructor(
-        address _asset,
-        address _config
-    )
+    constructor(address _asset, address _config)
         Ownable(msg.sender)
         ERC4626(IERC20(_asset))
-        ERC20(
-            string.concat("Vault ", IERC20Metadata(_asset).name()),
-            string.concat("v", IERC20Metadata(_asset).symbol())
-        )
+        ERC20(string.concat("Vault ", IERC20Metadata(_asset).name()), string.concat("v", IERC20Metadata(_asset).symbol()))
         EIP712(string.concat("Vault-", Strings.toHexString(_asset)), "1")
     {
         config = _config;
@@ -54,24 +49,16 @@ contract Vault is ERC4626, ERC20Votes, Ownable2Step, ReentrancyGuard {
         if (totalSupply() == NUMBER_OF_DEAD_SHARES) {
             require(msg.sender == owner(), "Only owner can call this function");
         } else {
-            require(
-                msg.sender == governor,
-                "Only governance can call this function"
-            );
+            require(msg.sender == governor, "Only governance can call this function");
         }
         _;
     }
 
-    function setWhitelisted(
-        address _contract,
-        bool _isWhitelisted
-    ) external onlyGovernanceOrOwner {
+    function setWhitelisted(address _contract, bool _isWhitelisted) external onlyGovernanceOrOwner {
         isWhitelisted[_contract] = _isWhitelisted;
     }
 
-    function setInterestRate(
-        uint _interestRatePerSecond
-    ) external onlyGovernanceOrOwner {
+    function setInterestRate(uint256 _interestRatePerSecond) external onlyGovernanceOrOwner {
         // TODO: should be delayed. or restricted with some conditions.
         lentAmountStored = lentAssets();
         lastUpdated = block.timestamp;
@@ -79,10 +66,7 @@ contract Vault is ERC4626, ERC20Votes, Ownable2Step, ReentrancyGuard {
     }
 
     modifier onlyWhitelisted(address _contract) {
-        require(
-            isWhitelisted[_contract],
-            "Only whitelisted contract can call this function"
-        );
+        require(isWhitelisted[_contract], "Only whitelisted contract can call this function");
         _;
     }
 
@@ -91,37 +75,23 @@ contract Vault is ERC4626, ERC20Votes, Ownable2Step, ReentrancyGuard {
     }
 
     function lentAssets() public view returns (uint256) {
-        uint duration = block.timestamp - lastUpdated;
-        return
-            InterestRate.calculatePrincipalPlusInterest(
-                lentAmountStored,
-                interestRatePerSecond,
-                duration
-            );
+        uint256 duration = block.timestamp - lastUpdated;
+        return InterestRate.calculatePrincipalPlusInterest(lentAmountStored, interestRatePerSecond, duration);
     }
 
     function _updateLentAmount(uint256 _amount, bool _add) private {
-        uint _lentAssets = lentAssets();
+        uint256 _lentAssets = lentAssets();
         lentAmountStored = _add ? _lentAssets + _amount : _lentAssets - _amount;
         lastUpdated = block.timestamp;
     }
 
-    function borrow(
-        uint256 _borrowAmount,
-        address _receiver
-    ) public nonReentrant onlyWhitelisted(msg.sender) {
+    function borrow(uint256 _borrowAmount, address _receiver) public nonReentrant onlyWhitelisted(msg.sender) {
         IERC20(asset()).safeTransfer(_receiver, _borrowAmount);
         _updateLentAmount(_borrowAmount, true);
     }
 
-    function repay(
-        uint256 _repayAmount
-    ) public nonReentrant onlyWhitelisted(msg.sender) {
-        IERC20(asset()).safeTransferFrom(
-            msg.sender,
-            address(this),
-            _repayAmount
-        );
+    function repay(uint256 _repayAmount) public nonReentrant onlyWhitelisted(msg.sender) {
+        IERC20(asset()).safeTransferFrom(msg.sender, address(this), _repayAmount);
         _updateLentAmount(_repayAmount, false);
     }
 
@@ -129,11 +99,7 @@ contract Vault is ERC4626, ERC20Votes, Ownable2Step, ReentrancyGuard {
         return uint48(block.number);
     }
 
-    function _update(
-        address from,
-        address to,
-        uint256 value
-    ) internal override(ERC20, ERC20Votes) {
+    function _update(address from, address to, uint256 value) internal override(ERC20, ERC20Votes) {
         super._update(from, to, value);
     }
