+++ test/foundry/Oracle.t.sol
@@ -9,15 +9,14 @@ import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {Oracle, PriceMessage} from "@oracle/Oracle.sol";
 
 contract OracleTest is Base {
-    
     // Additional test variables for EIP-712 specific tests
     address testKeeper = makeAddr("testKeeper");
     uint256 testKeeperPrivateKey = 0x1234;
     address testAsset = makeAddr("testAsset");
-    
+
     function setUp() public override {
         super.setUp();
-        
+
         // Set up additional keeper for EIP-712 tests
         vm.prank(deployer);
         oracle.setKeeper(vm.addr(testKeeperPrivateKey), true);
@@ -25,7 +24,7 @@ contract OracleTest is Base {
     }
 
     // ============ Basic Oracle Functionality Tests ============
-    
+
     function test_setKeeper() public {
         vm.startPrank(deployer);
         oracle.setKeeper(keeper1, true);
@@ -42,7 +41,7 @@ contract OracleTest is Base {
     function test_setHeartbeat() public {
         vm.startPrank(deployer);
         oracle.setHeartbeat(address(assets[0]), 1000);
-        (, , uint heartbeat) = oracle.referenceData(address(assets[0]));
+        (,, uint256 heartbeat) = oracle.referenceData(address(assets[0]));
         assertEq(heartbeat, 1000);
         vm.stopPrank();
     }
@@ -109,55 +108,55 @@ contract OracleTest is Base {
     }
 
     // ============ EIP-712 Signature Tests ============
-    
+
     function test_EIP712Signature() public {
         uint256 price = 1000e18;
-        
+
         // Create 3 signatures for the required minimum
         PriceMessage[] memory priceMessages = new PriceMessage[](3);
         priceMessages[0] = getPMsg(testAsset, price, testKeeperPrivateKey);
         priceMessages[1] = getPMsg(testAsset, price, keeper1Key);
         priceMessages[2] = getPMsg(testAsset, price, keeper2Key);
-        
+
         // Update price - should succeed with EIP-712 signature
         vm.prank(testKeeper);
         oracle.updatePrice(testAsset, priceMessages);
-        
+
         // Verify price was updated
-        (uint lastData, uint updatedTimestamp,) = oracle.referenceData(testAsset);
+        (uint256 lastData, uint256 updatedTimestamp,) = oracle.referenceData(testAsset);
         assertEq(lastData, price);
         assertEq(updatedTimestamp, block.timestamp);
     }
-    
+
     function test_EIP712DomainSeparation() public {
         uint256 price = 1000e18;
         uint256 chainId = block.chainid;
         uint256 timestamp = block.timestamp;
-        
+
         // Deploy a second oracle with different domain
         vm.prank(deployer);
         Oracle oracle2 = new Oracle(1);
         vm.prank(deployer);
         oracle2.setKeeper(testKeeper, true);
-        
+
         // Get hashes from both oracles
         bytes32 hash1 = oracle.getPriceMessageHash(testAsset, price, chainId, timestamp);
         bytes32 hash2 = oracle2.getPriceMessageHash(testAsset, price, chainId, timestamp);
-        
+
         // Hashes should be different due to domain separation
         assertNotEq(hash1, hash2, "Domain separation failed");
     }
-    
+
     function test_EIP712ReplayProtection() public {
         uint256 price = 1000e18;
         uint256 chainId = block.chainid;
         uint256 timestamp = block.timestamp;
-        
+
         // Create valid signature for current chain
         bytes32 digest = oracle.getPriceMessageHash(testAsset, price, chainId, timestamp);
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(testKeeperPrivateKey, digest);
         bytes memory signature = abi.encodePacked(r, s, v);
-        
+
         // Try to use signature with different chain ID (should fail)
         PriceMessage[] memory priceMessages = new PriceMessage[](3);
         priceMessages[0] = PriceMessage({
@@ -169,38 +168,38 @@ contract OracleTest is Base {
         });
         priceMessages[1] = getPMsg(testAsset, price, keeper1Key);
         priceMessages[2] = getPMsg(testAsset, price, keeper2Key);
-        
+
         vm.expectRevert("Oracle: invalid signer");
         vm.prank(testKeeper);
         oracle.updatePrice(testAsset, priceMessages);
     }
-    
+
     function test_EIP712TypeHashValidation() public {
         // Test that the type hash is correctly formed by verifying signature validation works
         // This indirectly tests that PRICE_MESSAGE_TYPEHASH matches the expected format
         uint256 price = 1000e18;
-        
+
         // Create 3 signatures for the required minimum
         PriceMessage[] memory priceMessages = new PriceMessage[](3);
         priceMessages[0] = getPMsg(testAsset, price, testKeeperPrivateKey);
         priceMessages[1] = getPMsg(testAsset, price, keeper1Key);
         priceMessages[2] = getPMsg(testAsset, price, keeper2Key);
-        
+
         // Should succeed, proving type hash is correct
         vm.prank(testKeeper);
         oracle.updatePrice(testAsset, priceMessages);
     }
-    
+
     function test_InvalidEIP712Signature() public {
         uint256 price = 1000e18;
         uint256 chainId = block.chainid;
         uint256 timestamp = block.timestamp;
-        
+
         // Create signature for different data
         bytes32 wrongDigest = oracle.getPriceMessageHash(testAsset, price + 1, chainId, timestamp);
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(testKeeperPrivateKey, wrongDigest);
         bytes memory signature = abi.encodePacked(r, s, v);
-        
+
         // Try to use signature with original data
         PriceMessage[] memory priceMessages = new PriceMessage[](3);
         priceMessages[0] = PriceMessage({
@@ -212,7 +211,7 @@ contract OracleTest is Base {
         });
         priceMessages[1] = getPMsg(testAsset, price, keeper1Key);
         priceMessages[2] = getPMsg(testAsset, price, keeper2Key);
-        
+
         vm.expectRevert("Oracle: invalid signer");
         vm.prank(testKeeper);
         oracle.updatePrice(testAsset, priceMessages);
@@ -244,7 +243,7 @@ contract OracleTest is Base {
     function test_DuplicateSigner() public {
         test_setKeeper();
         uint256 price = 1000e18;
-        
+
         // Use same signer twice
         PriceMessage[] memory pMsg = new PriceMessage[](3);
         pMsg[0] = getPMsg(testAsset, price, keeper1Key);
@@ -300,18 +299,18 @@ contract OracleTest is Base {
 
     function test_PauseUnpause() public {
         vm.startPrank(deployer);
-        
+
         // Pause oracle
         oracle.pause();
-        
+
         // Should revert when paused
         vm.expectRevert();
         oracle.priceOf(testAsset);
-        
+
         // Unpause oracle
         oracle.unpause();
         vm.stopPrank();
-        
+
         // Now should work (though no price set, will revert for different reason)
         vm.expectRevert("Oracle: price not updated");
         oracle.priceOf(testAsset);
